backend/requirements.txt
annotated-types==0.7.0
anyio==4.12.0
async-timeout==5.0.1
asyncpg==0.31.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
exceptiongroup==1.3.1
filelock==3.20.0
fsspec==2025.12.0
greenlet==3.3.0
grpcio==1.76.0
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
hyperframe==6.1.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==1.2.5
langchain-text-splitters==1.1.0
langsmith==0.6.0
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.4.2
nltk==3.9.2
numpy==2.2.6
orjson==3.11.5
packaging==25.0
portalocker==3.2.0
protobuf==6.33.2
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
python-crfsuite==0.9.12
python-dotenv==1.2.1
pytorch-triton-rocm==3.5.1
PyYAML==6.0.3
qdrant-client==1.16.2
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
sentence-transformers==5.2.0
shellingham==1.5.4
SQLAlchemy==2.0.45
sqlmodel==0.0.31
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.9.1
tqdm==4.67.1
transformers==4.57.3
typer-slim==0.21.0
typing-inspection==0.4.2
typing_extensions==4.15.0
underthesea==8.3.0
underthesea_core==1.0.5
urllib3==2.6.2
uuid_utils==0.12.0
zstandard==0.25.0
